CharacterAndFormat is now an interface
